refactor: removed nested components from the ResourceManagement.tsx (#500)
