+++ frontend/src/Pages/ResourcesManagement.tsx
@@ -3,8 +3,6 @@ import { AxiosError } from 'axios';
 import useSWR from 'swr';
 import Modal from '../Components/Modal';
 import AddResourceCollectionForm from '../Components/forms/AddResourceCollectionForm';
-import EditResourceCollectionForm from '../Components/forms/EditResourceCollectionForm';
-import AddLinkForm from '../Components/forms/AddLinkForm';
 import {
     ModalType,
     ResourceCategory,
@@ -12,16 +10,13 @@ import {
     ServerResponseMany,
     ToastState
 } from '../common';
-import { PencilSquareIcon, PlusCircleIcon } from '@heroicons/react/24/outline';
-import { TrashIcon } from '@heroicons/react/24/outline';
-import { Bars3Icon, PlusIcon } from '@heroicons/react/24/solid';
-import { CloudArrowUpIcon } from '@heroicons/react/16/solid';
+
+import { CloudArrowUpIcon, PlusIcon } from '@heroicons/react/16/solid';
 import DeleteForm from '../Components/forms/DeleteForm';
-import { useDebounceValue } from 'usehooks-ts';
-import ExternalLink from '@/Components/ExternalLink';
-import ULIComponent from '@/Components/ULIComponent.tsx';
 import { useToast } from '@/Context/ToastCtx.tsx';
 import API from '@/api/api';
+import SortableCollectionList from '@/Components/resourceMgtHelpers/SortableCollectionList';
+import ResourceCollectionEditor from '@/Components/resourceMgtHelpers/ResourceCollectionEditor';
 
 type EditableResourceCollection = ResourceCategory & {
     isModified: boolean;
@@ -352,438 +347,3 @@ export default function ResourcesManagement() {
         </div>
     );
 }
-
-const SortableCollectionList = ({
-    collections,
-    selectedCollectionIndex,
-    onResourceCollectionClick,
-    onDeleteCollectionClick,
-    onUpdateCollectionList
-}: {
-    collections: EditableResourceCollection[];
-    selectedCollectionIndex: number | undefined;
-    onResourceCollectionClick: (collection: EditableResourceCollection) => void;
-    onDeleteCollectionClick: (collectionId: number) => void;
-    onUpdateCollectionList: (newList: EditableResourceCollection[]) => void;
-}) => {
-    const [draggedItem, setDraggedItem] = useState<number>();
-    const [draggedOverItem, setDraggedOverItem] = useState<
-        undefined | number
-    >();
-    const dragOverItem = useDebounceValue(draggedOverItem, 100);
-
-    const handleSort = () => {
-        if (!draggedItem) return;
-
-        const insertAtIndex = dragOverItem;
-
-        // if dragged item is higher in the list, then should subtract a number from where it needs to be placed
-        if (draggedItem && draggedItem < dragOverItem[0]!) {
-            insertAtIndex[0] = insertAtIndex[0]! - 1;
-        }
-
-        const newCollectionList = [...collections];
-
-        //Remove the dragged item from the list, save its content.
-        const draggedItemContent = newCollectionList.splice(draggedItem, 1)[0];
-
-        // Insert the dragged item into its new position
-        if (insertAtIndex[0] === collections.length) {
-            newCollectionList.push(draggedItemContent);
-        } else {
-            newCollectionList.splice(insertAtIndex[0]!, 0, draggedItemContent);
-        }
-
-        // Check to see which collections had their positions modified.
-        const updatedCollectionList = newCollectionList.map(
-            (collection, index) => {
-                if (collection.id !== collections[index].id) {
-                    return {
-                        ...collection,
-                        isModified: true
-                    };
-                }
-                return collection;
-            }
-        );
-
-        //update the actual array
-        onUpdateCollectionList(updatedCollectionList);
-
-        setDraggedItem(undefined);
-        setDraggedOverItem(undefined);
-    };
-
-    const dragOver = (e: React.DragEvent<HTMLDivElement>, index: number) => {
-        e.preventDefault();
-        const source = e.dataTransfer.getData('text/plain');
-        if (source === 'collection-list') {
-            setDraggedOverItem(index);
-        }
-    };
-
-    const dragLeave = (e: React.DragEvent<HTMLDivElement>) => {
-        e.preventDefault();
-        setDraggedOverItem(undefined);
-    };
-
-    const dragDrop = (e: React.DragEvent<HTMLDivElement>) => {
-        e.preventDefault();
-        const source = e.dataTransfer.getData('text/plain');
-        if (source === 'collection-list') {
-            handleSort();
-        }
-    };
-
-    const dragStart = (e: React.DragEvent<HTMLDivElement>, index: number) => {
-        setDraggedItem(index);
-        e.dataTransfer.setData('text/plain', 'collection-list');
-    };
-
-    const dragEnd = (e: React.DragEvent<HTMLDivElement>) => {
-        e.preventDefault();
-        if (dragOverItem[0] == undefined) {
-            setDraggedOverItem(-1);
-        } else {
-            handleSort();
-        }
-    };
-
-    return collections.map((collection, index) => {
-        return (
-            <div
-                key={index}
-                className={`flex flex-col ${index === collections.length - 1 ? 'grow' : ''}`}
-            >
-                <div
-                    className={
-                        draggedItem == index
-                            ? dragOverItem[0] == -1
-                                ? 'block'
-                                : 'hidden'
-                            : 'block'
-                    }
-                >
-                    <div
-                        className={
-                            dragOverItem[0] == index
-                                ? 'pt-36 flex'
-                                : 'pt-2 flex'
-                        }
-                        onDragOver={(e) => dragOver(e, index)}
-                        onDragLeave={(e) => dragLeave(e)}
-                        onDrop={(e) => dragDrop(e)}
-                        onDragStart={(e) => dragStart(e, index)}
-                        onDragEnd={(e) => dragEnd(e)}
-                        draggable
-                    >
-                        <ResourceCollectionCardWithActions
-                            collection={collection}
-                            selected={index === selectedCollectionIndex}
-                            onResourceCollectionClick={
-                                onResourceCollectionClick
-                            }
-                            onDeleteCollectionClick={onDeleteCollectionClick}
-                        />
-                    </div>
-                </div>
-
-                {/* Render an empty div as a drop target for the end of the list. */}
-                {index === collections.length - 1 ? (
-                    <div
-                        className="grow"
-                        onDragOver={(e) => {
-                            e.preventDefault();
-                            setDraggedOverItem(index + 1);
-                        }}
-                        onDragLeave={(e) => {
-                            e.preventDefault();
-                            setDraggedOverItem(undefined);
-                        }}
-                        onDrop={(e) => {
-                            e.preventDefault();
-                            setDraggedItem(index);
-                        }}
-                    ></div>
-                ) : undefined}
-            </div>
-        );
-    });
-};
-
-const ResourceCollectionCardWithActions = ({
-    collection,
-    selected,
-    onResourceCollectionClick,
-    onDeleteCollectionClick
-}: {
-    collection: EditableResourceCollection;
-    selected: boolean;
-    onResourceCollectionClick: (collection: EditableResourceCollection) => void;
-    onDeleteCollectionClick: (collectionId: number) => void;
-}) => {
-    const cardClasses = `card card-compact grow overflow-visible ${collection.isModified ? 'bg-pale-yellow' : 'bg-inner-background'} ${selected ? 'border border-dark-yellow' : ''}`;
-
-    return (
-        <div
-            className={cardClasses}
-            onClick={() => onResourceCollectionClick(collection)}
-        >
-            <div className="card-body gap-2">
-                <div className="flex justify-between">
-                    <h3 className="card-title text-sm">{collection.name}</h3>
-                    <ULIComponent
-                        dataTip={'Delete Collection'}
-                        iconClassName={'self-start cursor-pointer'}
-                        onClick={() => onDeleteCollectionClick(collection.id)}
-                        icon={TrashIcon}
-                    />
-                </div>
-                {collection.links.map((link: ResourceLink, index: number) => {
-                    const [title, url] = Object.entries(link)[0];
-                    return (
-                        <ExternalLink key={index} url={url}>
-                            {title}
-                        </ExternalLink>
-                    );
-                })}
-            </div>
-        </div>
-    );
-};
-
-const ResourceCollectionEditor = ({
-    collection,
-    onCollectionChange,
-    onResourceLinkChange
-}: {
-    collection: EditableResourceCollection;
-    onCollectionChange: (updatedCollection: EditableResourceCollection) => void;
-    onResourceLinkChange: (
-        linkIndex: number,
-        updatedResourceLink: ResourceLink
-    ) => void;
-}) => {
-    const addLinkModal = useRef<HTMLDialogElement>(null);
-    const deleteLinkModal = useRef<HTMLDialogElement>(null);
-    const editResourceCollectionModal = useRef<HTMLDialogElement>(null);
-    const [activeLinkToDelete, setActiveLinkToDelete] = useState<
-        number | undefined
-    >();
-
-    const [draggedItem, setDraggedItem] = useState<number | undefined>();
-    const [draggedOverItem, setDraggedOverItem] = useState<
-        number | undefined
-    >();
-
-    const editCollectionTitle = (newtitle: string) => {
-        onCollectionChange({
-            ...collection,
-            name: newtitle,
-            isModified: true
-        });
-    };
-
-    const addLink = (
-        collection: EditableResourceCollection,
-        title: string,
-        url: string
-    ) => {
-        onCollectionChange({
-            ...collection,
-            links: [...collection.links, { [title]: url }],
-            isModified: true
-        });
-    };
-
-    const deleteLink = (
-        collection: EditableResourceCollection,
-        linkIndex: number
-    ) => {
-        onCollectionChange({
-            ...collection,
-            links: collection.links.filter((_, index) => index !== linkIndex),
-            isModified: true
-        });
-    };
-
-    const handleSort = () => {
-        if (!draggedItem || !draggedOverItem) return;
-
-        // Check to see if dragged item stayed in the same position
-        if (draggedItem === draggedOverItem) return;
-
-        const newLinks = [...collection.links];
-        const draggedLink = newLinks.splice(draggedItem, 1)[0];
-        newLinks.splice(draggedOverItem, 0, draggedLink);
-
-        onCollectionChange({
-            ...collection,
-            links: newLinks,
-            isModified: true
-        });
-
-        setDraggedItem(undefined);
-        setDraggedOverItem(undefined);
-    };
-
-    return (
-        <div className="card bg-inner-background">
-            <div className="card-body gap-2">
-                <div className="flex justify-between">
-                    <h3 className="card-title text-sm">{collection.name}</h3>
-                    <div className={'space-x-2'}>
-                        <ULIComponent
-                            iconClassName={'self-start cursor-pointer'}
-                            dataTip={'Edit Collection'}
-                            onClick={() =>
-                                editResourceCollectionModal.current?.showModal()
-                            }
-                            icon={PencilSquareIcon}
-                        />
-
-                        <ULIComponent
-                            iconClassName={'self-start cursor-pointer'}
-                            tooltipClassName={'self-start mr-2'}
-                            dataTip={'New Link'}
-                            onClick={() => addLinkModal.current?.showModal()}
-                            icon={PlusCircleIcon}
-                        />
-                    </div>
-                </div>
-                <table className="table">
-                    <thead>
-                        <tr className="border-grey-3">
-                            <th className="w-4"></th>
-                            <th>Link Name</th>
-                            <th>Link URL</th>
-                            <th className="w-4"></th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        {collection.links.map(
-                            (link: ResourceLink, linkIndex: number) => {
-                                const [title, url] = Object.entries(link)[0];
-                                return (
-                                    <tr
-                                        key={linkIndex}
-                                        draggable={draggedItem !== null}
-                                        onDragOver={(e) => {
-                                            e.preventDefault();
-                                            setDraggedOverItem(linkIndex);
-                                        }}
-                                        onDrop={handleSort}
-                                        onDragEnd={handleSort}
-                                        className={
-                                            draggedItem === linkIndex
-                                                ? 'bg-grey-1'
-                                                : ''
-                                        }
-                                    >
-                                        <td>
-                                            <ULIComponent
-                                                onMouseDown={() =>
-                                                    setDraggedItem(linkIndex)
-                                                }
-                                                icon={Bars3Icon}
-                                            />
-                                        </td>
-                                        <td>
-                                            <input
-                                                type="text"
-                                                className="input input-bordered w-full"
-                                                value={title}
-                                                onChange={(e) => {
-                                                    onResourceLinkChange(
-                                                        linkIndex,
-                                                        {
-                                                            [e.target.value]:
-                                                                url
-                                                        }
-                                                    );
-                                                }}
-                                            />
-                                        </td>
-                                        <td>
-                                            <input
-                                                type="text"
-                                                className="input input-bordered w-full"
-                                                value={url}
-                                                onChange={(e) =>
-                                                    onResourceLinkChange(
-                                                        linkIndex,
-                                                        {
-                                                            [title]:
-                                                                e.target.value
-                                                        }
-                                                    )
-                                                }
-                                            />
-                                        </td>
-                                        <td>
-                                            <ULIComponent
-                                                iconClassName={
-                                                    'self-start cursor-pointer'
-                                                }
-                                                dataTip={'Delete Link'}
-                                                onClick={() => {
-                                                    setActiveLinkToDelete(
-                                                        linkIndex
-                                                    );
-                                                    deleteLinkModal.current?.showModal();
-                                                }}
-                                                icon={TrashIcon}
-                                            />
-                                        </td>
-                                    </tr>
-                                );
-                            }
-                        )}
-                    </tbody>
-                </table>
-            </div>
-            <Modal
-                type={ModalType.Add}
-                item="Link"
-                form={
-                    <AddLinkForm
-                        onSuccess={(title: string, url: string) => {
-                            addLink(collection, title, url);
-                            addLinkModal.current?.close();
-                        }}
-                    />
-                }
-                ref={addLinkModal}
-            />
-            <Modal
-                type={ModalType.Confirm}
-                item="Delete Link"
-                form={
-                    <DeleteForm
-                        item="Link"
-                        onCancel={() => setActiveLinkToDelete(undefined)}
-                        onSuccess={() =>
-                            activeLinkToDelete != undefined &&
-                            deleteLink(collection, activeLinkToDelete)
-                        }
-                    />
-                }
-                ref={deleteLinkModal}
-            />
-            <Modal
-                type={ModalType.Edit}
-                item="Collection Name"
-                form={
-                    <EditResourceCollectionForm
-                        collectionName={collection.name}
-                        onSuccess={(newTitle: string) => {
-                            editCollectionTitle(newTitle);
-                            editResourceCollectionModal.current?.close();
-                        }}
-                    />
-                }
-                ref={editResourceCollectionModal}
-            />
-        </div>
-    );
-};
